#241 Refactor Court.vue — extract composables to reduce file size
Court.vue trimmed from 1,560 → 986 lines by extracting three composables:

- _useCourtComputeds.ts: all computed values (verdict, bench overview, tally display, verdict sections, justice targetings, bench label, preview preset, etc.)
- _useCourtHelpers.ts: UI helper functions (voteMeterStyle, verdictJusticeBarStyle, justiceVoteHistory) plus pure exports (getKeyStats, justiceTypeLabel, sideStanceTags, targetLabel, voteLabel, justiceImageUrl, presidentImageUrl, justiceHints)
- _useCourtTurns.ts: turn flow and opponent AI (endPlayerTurn, endOpponentTurn, playOpponentTurn, getEligibleMultiTargetJustices)
